refact: registerCars() 이동
Racing -> Cars
